textile-device-client/requirements.txt
requests>=2.31.0
psutil>=5.9.6
pystray>=0.19.5
Pillow>=10.1.0
PyQt6>=6.6.0
